desktop-sync/requirements.txt
telethon==1.36.0
PySide6==6.7.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1

